+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Fix issue with [23:10:26] [Wrapper.py/ERROR] Traceback (most recent call last):
 [23:10:26] [Wrapper.py/ERROR]   File "Wrapper.py/web.py", line 45, in onPlayerMessage
 [23:10:26] [Wrapper.py/ERROR]     del self.consoleScrollback[0]
 
-Stop backups from happening unless server is running. Handle running out of disk space by freezing java process. 
+Stop backups from happening unless server is running. Handle running out of disk space by freezing java process. Upon every boot, check level.dat and player files. If corrupted, replace from backup.
 <h4>0.7.5</h4>
 **Features**
 - Web interface improvements:
@@ -24,10 +24,12 @@ Stop backups from happening unless server is running. Handle running out of disk
 - Fixed crash if 'resource' module isn't installed (usually on non-POSIX systems)
 
 **Developer Changes**
+- api.registerHelp(groupName, summary, commands): Register new help menus to be displayed in the /help command. See documentation for more info
 - "AUTHOR" and "WEBSITE" plugin metadata variables added (see template.py for example)
 - player.say(message): Say something through the player. Proxy mode only.
 - player.execute(command): Execute a command as the player. Works best in proxy mode, but will fallback to using the 1.8 'execute' command if proxy mode is not available.
 - Fixed minecraft.getPlayer(username) so that it actually worked
+- server.getStorageAvailable(): Returns the amount of bytes free on the disk of the working directory
 
 <h4>0.7.4</h4>
 Just a small little update, to fix a few things, and improve upon some existing features.